stream_cdda: cast to int64_t before multiplication
Just to silence static analysis warning. It is assigned to int64_t
anyway.

+++ stream/stream_cdda.c
@@ -160,7 +160,7 @@ static int seek(stream_t *s, int64_t newpos)
     int seek_to_track = 0;
     int i;
 
-    newpos += p->start_sector * CDIO_CD_FRAMESIZE_RAW;
+    newpos += (int64_t)p->start_sector * CDIO_CD_FRAMESIZE_RAW;
 
     sec = newpos / CDIO_CD_FRAMESIZE_RAW;
     if (newpos < 0 || sec > p->end_sector) {
@@ -240,7 +240,7 @@ static int control(stream_t *stream, int cmd, void *arg)
 static int64_t get_size(stream_t *st)
 {
     cdda_priv *p = st->priv;
-    return (p->end_sector + 1 - p->start_sector) * CDIO_CD_FRAMESIZE_RAW;
+    return (int64_t)(p->end_sector + 1 - p->start_sector) * CDIO_CD_FRAMESIZE_RAW;
 }
 
 static int open_cdda(stream_t *st)
